Make it possible to handle multigraphs
We need to handle graphs which have multiple edges between the same pair
of vertices and which have edges that start and end in the same vertex.
The latter is currently broken because each vertex remembers a set of
edges which start in it and to handle single-edge loops, we need
a multiset there (with could continue using a set but it'd be more
complicated to implement `get_vertex_degree`). To prove that this now
works, we add a couple test cases of the Euler path algorithms which
relies on removing edges in multigraphs to work with no issues.