Add skeleton of the rest of the algorithm and tests
We add functions stubs for all the steps of the algorithm. Its interface
is now more or less complete so we also add a few tests which test cases
which we found difficult when designing it. The test are obviously not
passing yet so we mark all of them as "expected failure" (`xfail`).