Fix argument parsing to accept no arguments
When no arguments are given to the optimize.cli module on the command
line, it should default to reading from stdin. This is broken because
the `input_file`, as a positional argument, is required by default. We
fix this by specifying `nargs="?"` which makes it optional, with
a default value.