Change input format
We change the input format from

    <name> <x1> <y1> <x2> <y2> <C|O>

to:

    <name> <C|O> <x1> <y1> <x2> <y2>

because the interpretation of x/y values depends on whether the polyline
is closed or open. In fact, the input is a series of lines in one of the
two forms:

    <name> O <start_x> <start_y> <end_x> <end_y>
    <name> C <x_min> <y_min> <x_max> <y_max>

where the first type (`O`) are open polylines characterized by their
start and end, and the latter type (`C`) are closed polylines,
characterized by their bounding box.