Increase recursion limits during Euler path finding
We find Euler paths in graphs using a recursive algorithm which requires
the recursion depth to be as high as the number of edges in the graph.
The default recursion limit in Python is low (e.g., 1000) so to handle
bigger graphs we modify our `euler_path` algorithm to temporarily
increase it to the required value.